AST Tests: Print double constants at full precision.
This took care to leave most constants printed as they already were,
to reduce amount of test changes.

This will help test changes to strtod, as well as portability issues
with any current strtod.

The numbers printed do not look like they are always properly rounded,
so there may be issues with strtod beyond the locale "." vs. "," issue.
(Or, perhaps, issues with printf.)